# apigen/model/openapi.yaml
openapi: 3.0.3
info:
  title: Data Repository Service
  version: 1.5.0
  x-logo:
    url: https://www.ga4gh.org/wp-content/themes/ga4gh/dist/assets/svg/logos/logo-full-color.svg
  termsOfService: https://www.ga4gh.org/terms-and-conditions/
  contact:
    name: GA4GH Cloud Work Stream
    email: ga4gh-cloud@ga4gh.org
  license:
    name: Apache 2.0
    url: https://raw.githubusercontent.com/ga4gh/data-repository-service-schemas/master/LICENSE
servers:
  - url: https://{serverURL}/ga4gh/drs/v1
    variables:
      serverURL:
        default: drs.example.org
        description: |
          DRS server endpoints MUST be prefixed by the '/ga4gh/drs/v1' endpoint path
security:
  - {}
  - BasicAuth: []
  - BearerAuth: []
...
paths:
...
  /objects/{object_id}:
    options:
      summary: Get Authorization info about a DrsObject.
      security:
        - {}
      description: Returns a list of `Authorizations` that can be used to determine how to authorize requests to `GetObject` or `PostObject`.
      operationId: OptionsObject
      parameters:
        - $ref: '#/components/parameters/ObjectId'
      responses:
        '200':
          $ref: '#/components/responses/200OkAuthorizations'
        '204':
          $ref: '#/components/responses/AuthorizationsNotSupported'
        '400':
          $ref: '#/components/responses/400BadRequest'
        '404':
          $ref: '#/components/responses/404NotFoundDrsObject'
        '405':
          $ref: '#/components/responses/AuthorizationsNotSupported'
        '500':
          $ref: '#/components/responses/500InternalServerError'
      tags:
        - Objects
      x-swagger-router-controller: ga4gh.drs.server
    get:
      summary: Get info about a DrsObject.
      description: Returns object metadata, and a list of access methods that can be used to fetch object bytes.
      operationId: GetObject
      parameters:
        - $ref: '#/components/parameters/ObjectId'
        - $ref: '#/components/parameters/Expand'
      responses:
        '200':
          $ref: '#/components/responses/200OkDrsObject'
        '202':
          $ref: '#/components/responses/202Accepted'
        '400':
          $ref: '#/components/responses/400BadRequest'
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '403':
          $ref: '#/components/responses/403Forbidden'
        '404':
          $ref: '#/components/responses/404NotFoundDrsObject'
        '500':
          $ref: '#/components/responses/500InternalServerError'
      tags:
        - Objects
      x-swagger-router-controller: ga4gh.drs.server
    post:
      summary: Get info about a DrsObject through POST'ing a Passport.
      description: |-
        Returns object metadata and a list of access methods that can be used to fetch object bytes. Method is a POST to accommodate a JWT GA4GH Passport sent in the request body in order to authorize access.
        **Note**: To upload new files and register them as DRS objects, use the `/upload-request` endpoint  to obtain upload methods and temporary credentials, then use POST `/objects/register` endpoint  to register multiple objects at once. Note that upload functionality is optional and not all DRS  servers implement the upload endpoints.
      operationId: PostObject
      security:
        - PassportAuth: []
      responses:
        '200':
          $ref: '#/components/responses/200OkDrsObject'
        '202':
          $ref: '#/components/responses/202Accepted'
        '400':
          $ref: '#/components/responses/400BadRequest'
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '403':
          $ref: '#/components/responses/403Forbidden'
        '404':
          $ref: '#/components/responses/404NotFoundAccess'
        '500':
          $ref: '#/components/responses/500InternalServerError'
      tags:
        - Objects
      x-swagger-router-controller: ga4gh.drs.server
      parameters:
        - $ref: '#/components/parameters/ObjectId'
      requestBody:
        $ref: '#/components/requestBodies/PostObjectBody'
...
  /objects/register:
    post:
      summary: Register DRS objects
      description: "**Optional Endpoint**: This endpoint is not required for DRS server implementations.  Not all DRS servers support object registration.    \nRegisters one or more \"candidate\" DRS objects with the server. If it accepts the request, the server will create  unique object IDs for each registered object and return them in fully-formed DRS objects in response.\nThis endpoint can be used after uploading files using methods negotiated with the `/upload-request` endpoint  to register the uploaded files as DRS objects, or to register existing data. The request body should contain candidate  DRS objects with all required metadata including access methods that correspond  to the upload methods used during file upload.\n**RECOMMENDED - Transactional Behavior**:  Registration operations SHOULD be atomic transactions. If ANY candidate object fails validation  or registration, the ENTIRE request SHOULD fail and NO objects SHOULD be registered. Servers SHOULD  implement this as an all-or-nothing operation to ensure data consistency, but MAY implement partial  registration with appropriate error reporting if transactional behavior is not feasible.\n**Authentication**: GA4GH Passports can be provided in the request body for authorization. Bearer tokens can be supplied in headers.\n**Server Responsibilities**: - SHOULD treat registration as an atomic transaction (all succeed or all fail) - SHOULD validate ALL candidate objects before registering ANY - Create unique object IDs for each registered object - Add timestamps (created_time, updated_time) - SHOULD roll back any partial changes if any candidate fails validation\n**Client Responsibilities**: - Provide required DRS object metadata for all candidates - Include access methods corresponding to uploaded file locations - Ensure checksums match uploaded file content - Handle potential failure of entire batch if any single object is invalid"
      operationId: RegisterObjects
      security:
        - {}
        - BasicAuth: []
        - BearerAuth: []
        - PassportAuth: []
      requestBody:
        $ref: '#/components/requestBodies/RegisterObjectsBody'
      responses:
        '201':
          $ref: '#/components/responses/201ObjectsCreated'
        '400':
          $ref: '#/components/responses/400BadRequest'
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '403':
          $ref: '#/components/responses/403Forbidden'
        '413':
          $ref: '#/components/responses/413RequestTooLarge'
        '500':
          $ref: '#/components/responses/500InternalServerError'
      tags:
        - Register Objects
      x-swagger-router-controller: ga4gh.drs.server
      x-codegen-request-body-name: body
...
  /objects/checksum/{checksum}:
    get:
      summary: Get DRS objects that are a match for the checksum.
      description: |-
        Returns an array of `DRSObjects` that match a given checksum.
        The checksum type is not provide, the checksum check is done against all checksum types.
      operationId: GetObjectsByChecksum
      security:
        - PassportAuth: []
      parameters:
        - $ref: '#/components/parameters/Checksum'
      responses:
        '200':
          $ref: '#/components/responses/200OkDrsObjects'
        '202':
          $ref: '#/components/responses/202Accepted'
        '400':
          $ref: '#/components/responses/400BadRequest'
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '403':
          $ref: '#/components/responses/403Forbidden'
        '404':
          $ref: '#/components/responses/404NotFoundDrsObject'
        '500':
          $ref: '#/components/responses/500InternalServerError'
      tags:
        - Objects
      x-swagger-router-controller: ga4gh.drs.server
...
components:
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic
      description: |
        A valid authorization token must be passed in the 'Authorization' header,
        e.g. "Basic ${token_string}"
    BearerAuth:
      type: http
      scheme: bearer
      description: A valid authorization token must be passed in the 'Authorization' header, e.g. "Bearer ${token_string}"
    PassportAuth:
      type: http
      scheme: bearer
      x-in: body
      bearerFormat: JWT
      description: A valid GA4GH Passport must be passed in the body of an HTTP POST request as a tokens[] array.
  schemas:
...
    Error:
      type: object
      description: An object that can optionally include information about the error.
      properties:
        msg:
          type: string
          description: A detailed error message.
        status_code:
          type: integer
          description: The integer representing the HTTP status code (e.g. 200, 404).
...
    summary:
      type: object
      description: A summary of what was resolved.
      properties:
        requested:
          type: integer
          description: Number of items requested.
        resolved:
          type: integer
          description: Number of objects resolved.
        unresolved:
          type: integer
          description: Number of objects not resolved.
    unresolved:
      type: array
      description: Error codes for each unresolved drs objects.
      items:
        type: object
        properties:
          error_code:
            type: integer
          object_ids:
            type: array
            items:
              type: string
...
  responses:
...
    500InternalServerError:
      description: An unexpected error occurred.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    200OkDrsObject:
      description: The `DrsObject` was found successfully
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/DrsObject'
    202Accepted:
      description: |
        The operation is delayed and will continue asynchronously. The client should retry this same request after the delay specified by Retry-After header.
      headers:
        Retry-After:
          description: |
            Delay in seconds. The client should retry this same request after waiting for this duration. To simplify client response processing, this must be an integral relative time in seconds. This value SHOULD represent the minimum duration the client should wait before attempting the operation again with a reasonable expectation of success. When it is not feasible for the server to determine the actual expected delay, the server may return a brief, fixed value instead.
          schema:
            type: integer
            format: int64
    400BadRequest:
      description: The request is malformed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    401Unauthorized:
      description: The request is unauthorized.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    403Forbidden:
      description: The requester is not authorized to perform this action.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    404NotFoundDrsObject:
      description: The requested `DrsObject` wasn't found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    404NotFoundAccess:
      description: The requested `AccessURL` wasn't found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    200OkAuthorizations:
      description: '`Authorizations` were found successfully'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Authorizations'
    AuthorizationsNotSupported:
      description: '`Authorizations` are not supported for this object. Default to `None`.'
    200OkDrsObjects:
      description: The `DrsObjects` were found successfully
      content:
        application/json:
          schema:
            type: object
            properties:
              summary:
                $ref: '#/components/schemas/summary'
              unresolved_drs_objects:
                $ref: '#/components/schemas/unresolved'
              resolved_drs_object:
                type: array
                items:
                  $ref: '#/components/schemas/DrsObject'
    413RequestTooLarge:
      description: The bulk request is too large.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            bulk_limit_exceeded:
              summary: Bulk delete limit exceeded
              description: Request contains more objects than server's maximum bulk delete limit
              value:
                msg: Bulk delete request contains 150 objects but server maximum is 100. Check maxBulkDeleteLength in service-info.
                status_code: 413
            request_size_too_large:
              summary: Request payload too large
              description: The overall request payload exceeds server limits
              value:
                msg: Request payload size exceeds server limit of 1MB
                status_code: 413
...
  parameters:
    ObjectId:
      in: path
      name: object_id
      required: true
      description: '`DrsObject` identifier'
      schema:
        type: string
    Expand:
      in: query
      name: expand
      schema:
        type: boolean
      example: false
      description: |-
        If false and the object_id refers to a bundle, then the ContentsObject array contains only those objects directly contained in the bundle. That is, if the bundle contains other bundles, those other bundles are not recursively included in the result.
        If true and the object_id refers to a bundle, then the entire set of objects in the bundle is expanded. That is, if the bundle contains other bundles, then those other bundles are recursively expanded and included in the result. Recursion continues through the entire sub-tree of the bundle.
        If the object_id refers to a blob, then the query parameter is ignored.
...
    Checksum:
      in: path
      name: checksum
      required: true
      description: A `checksum` value from the `checksums` list of a `DrsObject`
      schema:
        type: string
      x-go-name: ChecksumParam
  requestBodies:
    PostObjectBody:
      required: true
      content:
        application/json:
          schema:
            type: object
            properties:
              expand:
                type: boolean
                example: false
                description: |-
                  If false and the object_id refers to a bundle, then the ContentsObject array contains only those objects directly contained in the bundle. That is, if the bundle contains other bundles, those other bundles are not recursively included in the result.
                  If true and the object_id refers to a bundle, then the entire set of objects in the bundle is expanded. That is, if the bundle contains other bundles, then those other bundles are recursively expanded and included in the result. Recursion continues through the entire sub-tree of the bundle.
                  If the object_id refers to a blob, then the query parameter is ignored.
              passports:
                type: array
                items:
                  type: string
                  example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJnYTRnaF9wYXNzcG9ydF92MSI6W119.JJ5rN0ktP0qwyZmIPpxmF_p7JsxAZH6L6brUxtad3CM
                description: the encoded JWT GA4GH Passport that contains embedded Visas.  The overall JWT is signed as are the individual Passport Visas.
          examples:
            retrieve_with_auth:
              summary: Retrieve object with authentication
              description: Request object metadata with passport authentication
              value:
                expand: false
                passports:
                  - eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJnYTRnaF9wYXNzcG9ydF92MSI6W119.JJ5rN0ktP0qwyZmIPpxmF_p7JsxAZH6L6brUxtad3CM
            retrieve_expanded_bundle:
              summary: Retrieve expanded bundle with authentication
              description: Request expanded bundle contents with passport authentication
              value:
                expand: true
                passports:
                  - eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJnYTRnaF9wYXNzcG9ydF92MSI6W119.JJ5rN0ktP0qwyZmIPpxmF_p7JsxAZH6L6brUxtad3CM
                  - eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9.eyJnYTRnaF9wYXNzcG9ydF92MSI6W119.additional_passport_signature
...
    RegisterObjectsBody:
      description: Request body for registering DRS objects after upload
      required: true
      content:
        application/json:
          schema:
            type: object
            required:
              - candidates
            properties:
              candidates:
                type: array
                items:
                  $ref: '#/components/schemas/DrsObjectCandidate'
                minItems: 1
                description: Array of DRS object candidates to register (server will mint IDs and timestamps)
              passports:
                type: array
                items:
                  type: string
                description: Optional array of GA4GH Passport JWTs for authorization
          examples:
            single_object_registration:
              summary: Register a single object
              description: Register one DRS object after upload
              value:
                candidates:
                  - name: sample_data.vcf
                    size: 1048576
                    mime_type: text/plain
                    checksums:
                      - checksum: e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855
                        type: sha-256
                    description: Variant call format file for sample analysis
                    access_methods:
                      - type: s3
                        access_url:
                          url: s3://my-bucket/uploads/sample_data.vcf
                passports:
                  - eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
            bulk_object_registration:
              summary: Register multiple objects
              description: Register multiple DRS objects in a single request
              value:
                candidates:
                  - name: genome_assembly.fasta
                    size: 3221225472
                    mime_type: text/plain
                    checksums:
                      - checksum: a665a45920422f9d417e4867efdc4fb8a04a1f3fff1fa07e998e86f7f7a27ae3
                        type: sha-256
                    description: Human genome reference assembly
                    access_methods:
                      - type: s3
                        access_url:
                          url: s3://genomics-bucket/assemblies/hg38.fasta
                  - name: annotations.gff3
                    size: 524288000
                    mime_type: text/plain
                    checksums:
                      - checksum: b94d27b9934d3e08a52e52d7da7dabfac484efe37a5380ee9088f7ace2efcde9
                        type: sha-256
                    description: Gene annotations in GFF3 format
                    access_methods:
                      - type: https
                        access_url:
                          url: https://data.example.org/files/annotations.gff3
              controlled_access_registration:
                summary: Register a controlled-access object
                description: Register one DRS object and explicitly assign controlled-access claims. The server validates the supplied Passport and verifies that the caller is authorized to assign the requested claims.
                value:
                  candidates:
                    - name: tumor-normal.cram
                      size: 4294967296
                      mime_type: application/octet-stream
                      checksums:
                        - checksum: 9f86d081884c7d659a2feaa0c55ad015a3bf4f1b2b0b822cd15d6c15b0f00a08
                          type: sha-256
                      description: Controlled-access tumor-normal CRAM
                      controlled_access:
                        - https://auth.example.org/datasets/phs000123
                        - https://auth.example.org/projects/lung-study
                      access_methods:
                        - type: s3
                          access_url:
                            url: s3://controlled-bucket/phs000123/tumor-normal.cram
                  passports:
                    - eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9...
              open_access_registration:
                summary: Register an open-access object
                description: Register one DRS object without controlled-access claims. Because `controlled_access` is missing, the object is treated as open access for read operations.
                value:
                  candidates:
                    - name: public-reference.fa
                      size: 1024
                      mime_type: text/plain
                      checksums:
                        - checksum: 2c26b46b68ffc68ff99b453c1d30413413422d706483bfa0f98a5e886266e7ae
                          type: sha-256
                      description: Open-access public reference file
                      access_methods:
                        - type: https
                          access_url:
                            url: https://data.example.org/public/public-reference.fa
...